requirements.txt
transformers==4.16.2
torchvision==0.11.3
statsmodels==0.13.2
nltk==3.7
scikit-learn==1.0.2
modAL==0.4.1
datasets==1.18.3
pandas~=1.3.5
numpy~=1.21.5
# line for installing torch/torchvision if using CUDA to build model - Laptop GTX 3080. This will differ depending on your GPU.
#torch==1.9.0+cu111 torchvision==0.10.0+cu111 -f https://download.pytorch.org/whl/torch_stable.html




